pyproject.toml
[project]
name = "reviewweb"
version = "0.1.0"
description = ""
authors = [
    {name = "monda95",email = "jungmonda@naver.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
